Dedent only the raw array, and cook the dedented raw
Fixes #57.

This changes the algorithm so that we only dedent the raw array, meaning that any escape sequences are treated as non-whitespace chars for the calculate step. We then cook the raw string into a cooked string for the cooked array.

The practical change is demonstrated as:

```js
// A)
String.dedent`
  \x20 foo
`;

// B)
String.dedent`
  foo\n  bar
`
```

A) now prints a cooked `"··foo"` instead of `"foo"`. The raw remains `"\\x20·foo"`.
B) now prints a cooked `"foo\n··bar"` instead of `"foo\nbar"`. The raw remains `"foo\\n··bar"`.